Allow Kue configuration
The creation of a new Kue instance is done without passing any
configuration options, which makes impossible for the developer to
customize the created queue (for example: use unix socket instead of
a TCP/IP connection). Furthermore, Kueue uses a singleton for the queue,
so, creating a KueCleanup instance before any other invocation to
"createQueue" could result in Kue ignoring any configuration parameters
that the developer assumed that where going to take effect when
making those invocations.

Add a "config" parameter to the KueCleanup constructor that is passed
to the createQueue method, in order to configure the queue that is
created.